moved stock in history to display page
